+++ packages/sv-utils/src/index.ts
@@ -67,7 +67,7 @@ export const parse = {
 };
 
 // Utilities
-export { splitVersion, isVersionUnsupportedBelow } from './common.ts';
+export { splitVersion, coerceVersion, isVersionUnsupportedBelow, minVersion } from './semver.ts';
 export { createPrinter } from './utils.ts';
 export { sanitizeName } from './sanitize.ts';
 export { downloadJson } from './downloadJson.ts';